requirements.txt
aioredis==1.0.0
amqp==1.4.9
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
billiard==3.3.0.23
celery==3.1.18
certifi==2018.1.18
channels==1.1.8
channels-redis==2.1.1
chardet==3.0.4
constantly==15.1.0
daphne==1.4.2
Django==2.0.3
django-simple-history==1.9.0
future==0.16.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
kombu==3.0.37
msgpack==0.5.6
msgpack-python==0.5.6
oauthlib==2.0.7
Pillow==5.0.0
python-twitter==3.4.1
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
websocket-client==0.47.0
zope.interface==4.4.3
